Move cli-related files to another directory
